fix datepicker minDate and maxDate
