+++ src/components/DatePicker/index.jsx
@@ -1,43 +1,25 @@
-import React, { forwardRef, useState } from "react";
+import React, { forwardRef, useState, useEffect } from "react";
 
-import { DatePicker as AntDatePicker, ConfigProvider } from "antd";
+import { DatePicker as AntDatePicker } from "antd";
 import classnames from "classnames";
+import { isNotPresent } from "neetocist";
 import { Left, Right, Calendar, Close } from "neetoicons";
 import PropTypes from "prop-types";
 
+import Label from "components/Label";
 import { useSyncedRef, useId } from "hooks";
-import {
-  convertToDayjsObjects,
-  noop,
-  hyphenize,
-  ANT_DESIGN_GLOBAL_TOKEN_OVERRIDES,
-} from "utils";
+import { convertToDayjsObjects, noop, hyphenize } from "utils";
 
-import Label from "./Label";
+import IconOverride from "./IconOverride";
+import Provider from "./Provider";
+import Today from "./Today";
+import { getAllowed } from "./utils";
 
 const INPUT_SIZES = { small: "small", medium: "medium", large: "large" };
 
 const { RangePicker } = AntDatePicker;
 
 const datePickerTypes = { range: RangePicker, date: AntDatePicker };
-
-const IconOverride = ({ icon: Icon }) => (
-  <span className="neeto-ui-btn neeto-ui-btn--style-text neeto-ui-btn--size-medium neeto-ui-btn--icon-only">
-    <Icon className="neeto-ui-btn__icon" size={20} />
-  </span>
-);
-
-const Today = ({ onClick }) => (
-  <div className="text-center">
-    <button
-      {...{ onClick }}
-      className="neeto-ui-rounded-md hover:neeto-ui-bg-gray-200 px-2 py-1 text-xs font-medium transition duration-300 ease-in-out"
-      data-cy="year-month-mode-today"
-    >
-      Today
-    </button>
-  </div>
-);
 
 const DatePicker = forwardRef(
   (
@@ -58,14 +40,17 @@ const DatePicker = forwardRef(
       nakedInput = false,
       error = "",
       defaultValue,
-      value,
+      value: inputValue,
       labelProps,
       required = false,
       allowClear = true,
+      maxDate,
+      minDate,
       ...otherProps
     },
     ref
   ) => {
+    const [value, setValue] = useState(inputValue);
     const [mode, setMode] = useState(inputMode);
     const [pickerValue, setPickerValue] = useState();
     const id = useId(otherProps.id);
@@ -76,10 +61,17 @@ const DatePicker = forwardRef(
 
     const errorId = `error_${id}`;
 
+    useEffect(() => {
+      setValue(inputValue);
+    }, [inputValue]);
+
     const handleOnChange = (date, dateString) => {
-      type === "range" && !date
-        ? onChange([], dateString)
-        : onChange(date, dateString);
+      if (type == "range" && isNotPresent(date)) {
+        return onChange([], dateString);
+      }
+      const allowed = getAllowed(date, minDate, maxDate);
+      setValue(allowed);
+      onChange(allowed, dateString);
     };
 
     const renderExtraFooter = () => {
@@ -98,60 +90,7 @@ const DatePicker = forwardRef(
     };
 
     return (
-      <ConfigProvider
-        theme={{
-          token: { ...ANT_DESIGN_GLOBAL_TOKEN_OVERRIDES },
-          components: {
-            DatePicker: {
-              activeBg: "rgb(var(--neeto-ui-white))",
-              activeBorderColor: "rgb(var(--neeto-ui-primary-500))",
-              addonBg: "rgb(var(--neeto-ui-gray-100))",
-              cellActiveWithRangeBg: "rgb(var(--neeto-ui-primary-100))",
-              cellBgDisabled: "rgb(var(--neeto-ui-gray-100))",
-              cellHoverBg: "rgb(var(--neeto-ui-gray-200))",
-              cellHoverWithRangeBg: "rgb(var(--neeto-ui-primary-100))",
-              cellRangeBorderColor: "rgb(var(--neeto-ui-primary-100))",
-              hoverBg: "rgb(var(--neeto-ui-white))",
-              hoverBorderColor: "rgb(var(--neeto-ui-primary-500))",
-
-              // Global overrides
-              colorBgContainer: "rgb(var(--neeto-ui-white))",
-              colorBgElevated: "rgb(var(--neeto-ui-white))",
-              colorPrimary: "rgb(var(--neeto-ui-primary-500))",
-              colorPrimaryBorder: "rgb(var(--neeto-ui-primary-100))",
-              colorPrimaryHover: "rgb(var(--neeto-ui-primary-600))",
-              colorBorder: "rgb(var(--neeto-ui-gray-300))",
-              colorError: "rgb(var(--neeto-ui-error-500))",
-              colorErrorHover: "rgb(var(--neeto-ui-error-600))",
-              colorErrorOutline: "rgb(var(--neeto-ui-error-100))",
-              colorFillAlter: "rgb(var(--neeto-ui-gray-100))",
-              colorIcon: "rgb(var(--neeto-ui-gray-700))",
-              colorIconHover: "rgb(var(--neeto-ui-gray-800))",
-              colorLink: "rgb(var(--neeto-ui-primary-500))",
-              colorLinkHover: "rgb(var(--neeto-ui-primary-600))",
-              colorLinkActive: "rgb(var(--neeto-ui-primary-800))",
-              colorSplit: "rgb(var(--neeto-ui-gray-200))",
-              colorText: "rgb(var(--neeto-ui-gray-800))",
-              colorTextDescription: "rgb(var(--neeto-ui-gray-700))",
-              colorTextDisabled: "rgb(var(--neeto-ui-gray-500))",
-              colorTextHeading: "rgb(var(--neeto-ui-black))",
-              colorTextLightSolid: "rgb(var(--neeto-ui-white))",
-              colorTextPlaceholder: "rgb(var(--neeto-ui-gray-400))",
-              colorTextQuaternary: "rgb(var(--neeto-ui-gray-400))",
-              colorWarning: "rgb(var(--neeto-ui-warning-500))",
-              colorWarningHover: "rgb(var(--neeto-ui-warning-600))",
-              colorWarningOutline: "rgb(var(--neeto-ui-warning-100))",
-              controlItemBgActive: "rgb(var(--neeto-ui-pastel-purple))",
-              controlItemBgHover: "rgb(var(--neeto-ui-gray-100))",
-              controlOutline: "rgb(var(--neeto-ui-gray-300))",
-
-              // Sizes
-              cellHeight: 32,
-              padding: 22,
-            },
-          },
-        }}
-      >
+      <Provider>
         <div className="neeto-ui-input__wrapper">
           {label && <Label {...{ required, ...labelProps }}>{label}</Label>}
           <Component
@@ -175,6 +114,8 @@ const DatePicker = forwardRef(
             onChange={handleOnChange}
             {...{
               format,
+              maxDate,
+              minDate,
               onOk,
               picker,
               ...otherProps,
@@ -209,7 +150,7 @@ const DatePicker = forwardRef(
             </p>
           )}
         </div>
-      </ConfigProvider>
+      </Provider>
     );
   }
 );
